+++ src/core/__spec__/flatten.mock.ts
@@ -1,4 +1,4 @@
-import type { SplitPanelDef } from './interfaces';
+import type { SplitPanelDef } from '../interfaces';
 
 export const examplePanel1: SplitPanelDef = {
   id: 'test1',
